factored file logging factory into it's own module
