+++ vlab_ipam_api/lib/__init__.py
@@ -2,3 +2,4 @@
 from .constants import const
 from .database import Database, DatabaseError
 from .firewall import FireWall
+from .file_logger import get_logger
